fix: move sample_data_fuse.py into the model fuser
fuse_sample_data (native-transform + predictor sample-data merge) is
consumed only by assembler tasks fusing a native-transform model, the
same fuser-consumer pattern as fuse_schema.py. Move it out of
lib/model_manager/schema/ into lib/shared/utils/model_fuser/ alongside
the rest of the fuser, and re-export it from the fuser's __init__.
