+++ python/michelangelo/lib/shared/utils/model_fuser/__init__.py
@@ -16,6 +16,7 @@ from .fuse import (
 )
 from .fuse_schema import fuse_input_schema, fuse_model_schema
 from .fused_model import FusedModel
+from .sample_data_fuse import fuse_sample_data
 
 __all__ = [
     "FusedModel",
@@ -26,5 +27,6 @@ __all__ = [
     "fuse_models_to_onnx",
     "fuse_models_to_python",
     "fuse_models_to_torchscript",
+    "fuse_sample_data",
     "get_predictor_output_field_order",
 ]
